Move contact scopes types to CampaignRecipients endpoint namespace
