+++ src/endpoints/CampaignRecipients/index.ts
@@ -1,2 +1,3 @@
 export * from './Client';
+export * from './scopes';
 export * from './types';
