--- src/types/scopes/AllContactsScope.ts
@@ -1,7 +1,0 @@
-import { SelectionValue } from '../common';
-import { Contact } from '../Contact';
-
-export interface AllContactsScope {
-    type: 'scope:contacts';
-    selection?: SelectionValue<Contact['id']>;
-}
