--- src/types/scopes/ContactEmployeesScope.ts
@@ -1,8 +1,0 @@
-import { SelectionValue } from '../common';
-import { Contact } from '../Contact';
-
-export interface ContactEmployeesScope {
-    type: 'scope:contact_organisations';
-    contact_id: Contact['id'];
-    selection?: SelectionValue<Contact['id']>;
-}
