--- src/types/scopes/index.ts
@@ -1,6 +1,0 @@
-export * from './AllContactsScope';
-export * from './CampaignRecipientsScope';
-export * from './CampaignReportScope';
-export * from './ContactEmployeesScope';
-export * from './ContactOrganisationsScope';
-export * from './ContactsScope';
